requirements.txt
asgiref==3.2.10
boto3==1.14.47
botocore==1.17.47
dj-database-url==0.5.0
Django==3.1
django-appconf==1.0.4
django-compressor==2.4
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-libsass==0.8
django-sass-processor==0.8
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
libsass==0.20.0
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
rcssmin==1.0.6
rjsmin==1.1.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
